+++ opentelemetry-sdk/src/opentelemetry/sdk/_events/__init__.py
@@ -13,26 +13,31 @@
 # limitations under the License.
 
 import logging
-import warnings
 from time import time_ns
 from typing import Optional
+
+from typing_extensions import deprecated
 
 from opentelemetry import trace
 from opentelemetry._events import Event
 from opentelemetry._events import EventLogger as APIEventLogger
 from opentelemetry._events import EventLoggerProvider as APIEventLoggerProvider
-from opentelemetry._logs import NoOpLogger, SeverityNumber, get_logger_provider
-from opentelemetry.sdk._logs import (
-    LogDeprecatedInitWarning,
-    Logger,
-    LoggerProvider,
+from opentelemetry._logs import (
     LogRecord,
+    NoOpLogger,
+    SeverityNumber,
+    get_logger_provider,
 )
+from opentelemetry.sdk._logs import Logger, LoggerProvider
 from opentelemetry.util.types import _ExtendedAttributes
 
 _logger = logging.getLogger(__name__)
 
 
+@deprecated(
+    "You should use `Logger` instead. "
+    "Deprecated since version 1.39.0 and will be removed in a future release."
+)
 class EventLogger(APIEventLogger):
     def __init__(
         self,
@@ -58,25 +63,24 @@ class EventLogger(APIEventLogger):
             return
         span_context = trace.get_current_span().get_span_context()
 
-        # silence deprecation warnings from internal users
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", category=LogDeprecatedInitWarning)
-
-            log_record = LogRecord(
-                timestamp=event.timestamp or time_ns(),
-                observed_timestamp=None,
-                trace_id=event.trace_id or span_context.trace_id,
-                span_id=event.span_id or span_context.span_id,
-                trace_flags=event.trace_flags or span_context.trace_flags,
-                severity_text=None,
-                severity_number=event.severity_number or SeverityNumber.INFO,
-                body=event.body,
-                resource=getattr(self._logger, "resource", None),
-                attributes=event.attributes,
-            )
+        log_record = LogRecord(
+            timestamp=event.timestamp or time_ns(),
+            observed_timestamp=None,
+            trace_id=event.trace_id or span_context.trace_id,
+            span_id=event.span_id or span_context.span_id,
+            trace_flags=event.trace_flags or span_context.trace_flags,
+            severity_text=None,
+            severity_number=event.severity_number or SeverityNumber.INFO,
+            body=event.body,
+            attributes=event.attributes,
+        )
         self._logger.emit(log_record)
 
 
+@deprecated(
+    "You should use `LoggerProvider` instead. "
+    "Deprecated since version 1.39.0 and will be removed in a future release."
+)
 class EventLoggerProvider(APIEventLoggerProvider):
     def __init__(self, logger_provider: Optional[LoggerProvider] = None):
         self._logger_provider = logger_provider or get_logger_provider()